Improve install.sh

+++ install.sh
@@ -1,38 +1,38 @@
 #! /bin/bash
 
 dotfiles=~/dotfiles
-files="vimrc gitconfig" #Liste de tous les fichiers à déplacer
+files="vimrc gitconfig" 
 vim_theme="dark.vim light.vim"
-
-# Vim
-echo "Création du dossier vim"
-mkdir ~/.vim
-mkdir ~/.vim/colors
-
-echo "Déplacement vers le dossier $dotfiles"
-cd $dotfiles
 
 #Création des symlinks vers ~
 for file in $files; do
+    if [ -e "$HOME/$file" ]; then
     echo "Création du symlink de $file vers ~"
     ln -s $dotfiles/$file ~/.$file
+    fi
 done
 
-#Création fichier colors
-echo "Déplacement vers le dossier .vim"
-cd ~/.vim
-
-#Retour dans le dossier dotfiles et création des symlinks vers .vim/colors
-cd $dotfiles
+# Vim
+if [ ! -d "$HOME/.vim/colors" ]; then
+    echo "Création du dossier vim et vim/colors"
+    mkdir -p ~/.vim/colors
+fi
 for theme in $vim_theme; do
-    echo "Création du symlink de $vim_theme vers ~/.vim/colors"
-    ln -s $dotfiles/$theme ~/.vim/colors/$theme
+    if [ ! -e "$HOME/.vim/colors/$theme" ]; then
+        echo "Création du symlink de $theme vers ~/.vim/colors"
+        ln -s $dotfiles/$theme ~/.vim/colors/$theme
+    fi
 done
 
 #Création du fichier .zshrc
-touch ~/.zshrc
-echo "source ~/dotfiles/zshrc_config.zsh" >> ~/.zshrc
-echo "source ~/dotfiles/aliases" >> ~/.zshrc
-echo "source ~/dotfiles/minimal.zsh-theme" >> ~/.zshrc
-echo "source ~/dotfiles/zsh-autosuggestions/zsh-autosuggestions.zsh" >> ~/.zshrc
-echo "source ~/dotfiles/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> ~/.zshrc
+if [ ! -f "$HOME/.zshrc" ]; then
+    echo "Création du fichier .zshrc et ajout des 'sources'"
+    touch ~/.zshrc
+    echo "source ~/dotfiles/zshrc_config.zsh" >> ~/.zshrc
+    echo "source ~/dotfiles/aliases" >> ~/.zshrc
+    echo "source ~/dotfiles/minimal.zsh-theme" >> ~/.zshrc
+    echo "source ~/dotfiles/zsh-autosuggestions/zsh-autosuggestions.zsh" >> ~/.zshrc
+    echo "source ~/dotfiles/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh" >> ~/.zshrc
+fi
+
+echo "Setup completed!"
